pyproject.toml
[tool.poetry]
name = "caching-service"
version = "0.1.0"
description = ""
authors = ["Khaidzir Muhammad Shahih"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.4"}


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.24.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
